Saving data into Class FeaturesClass is almost completed.
The important Lesson here is that one can query one's own data by many
kinds of method such as list, hashmap, etc.

But should remind oneself that though admitting hard coding is not a
good way of programming, but it doesn't mean that it's totally
unnecessary in programming. 

More important thing in coding is not the efficiency of the skill or
technique itself but the way it works in proper way in proper coding
way.

Understanding the structure of data management makes it easier.